requirements.txt
fastapi
uvicorn
jinja2
openai
python-dotenv
pydantic
